+++ src/stories/icon.stories.mdx
@@ -1,4 +1,5 @@
 import { Canvas, Meta, Story } from '@storybook/addon-docs';
+import { paths } from '../elements/icon/icons'
 
 <Meta
   title='Elements/Icon'
@@ -39,6 +40,7 @@ import { Canvas, Meta, Story } from '@storybook/addon-docs';
         size='${size}'
         fill='${fill}'
       />
+      ${paths}
     `}
   </Story>
 </Canvas>
